+++ CHANGES.md
@@ -1,3 +1,10 @@
+
+* `Logs_fmt.reporter`, the optional argument `prefix` is changed
+  to `header` and becomes a formatter.
+* Fix colors in `Logs_fmt.pp_header`, only `Logs.err_style` was
+  begin used.
+
+
 v0.5.0 2016-01-07 La Forclaz (VS)
 ---------------------------------